refactor: 마이페이지(이미지 수정) 훅 합치기
