+++ src/components/UploadImage/index.jsx
@@ -1,3 +1,4 @@
+import { useRecoilState } from 'recoil';
 import UploadIcon from '../../../static/images/upload.svg';
 import { ImageFileInput } from '../AddPost/AddPostForm/style';
 import {
@@ -7,10 +8,30 @@ import {
   UploadDescription,
   UploadImageWrapper,
 } from './style';
-import { useMyhomeUploadFile } from '../../hooks/useMyhomeUploadFile';
 
-const UploadAndDisplayImage = ({ postId }) => {
-  const { selectedImage, handleImageChange } = useMyhomeUploadFile();
+import { useCallback } from 'react';
+import { uploadImageState } from '../../recoil/uploadImageState';
+import { ERROR_MESSAGE, FILE } from '../../utils/constants/myHome';
+
+const UploadAndDisplayImage = () => {
+  const [selectedImage, setSelectedImage] = useRecoilState(uploadImageState);
+
+  const checkImage = (file) => {
+    if (file.type === FILE.JPEG || file.type === FILE.SVG || file.type === FILE.PNG) {
+      setSelectedImage(file);
+    } else {
+      swal(ERROR_MESSAGE.UPLOAD_IMAGE);
+      return;
+    }
+  };
+
+  const handleImageChange = useCallback(
+    (e) => {
+      const file = e.target.files[0];
+      checkImage(file);
+    },
+    [selectedImage]
+  );
 
   return (
     <ImageUploaderContainer>
@@ -39,7 +60,6 @@ const UploadAndDisplayImage = ({ postId }) => {
                 width: '300px',
               }}
             >
-              {/*{image && <Image src={image} alt="사진" />}*/}
               <UploadDescription>Supported formates: JPEG,PNG,SVG</UploadDescription>
               <ImageFileInput type="file" onChange={handleImageChange} />
             </div>
